LayoutTests/imported/mozilla/css-animations/test_cssanimation-animationname.html

<!doctype html><!-- webkit-test-runner [ enableCSSAnimationsAndCSSTransitionsBackedByWebAnimations=true ] -->
<meta charset=utf-8>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../resources/testcommon.js"></script>
<style>
@keyframes xyz {
  to { left: 100px }
}
</style>
<body>
<div id="log"></div>
<script>
'use strict';

test(function(t) {
  var div = addDiv(t);
  div.style.animation = 'xyz 100s';
  assert_equals(div.getAnimations()[0].animationName, 'xyz',
                'Animation name matches keyframes rule name');
}, 'Animation name makes keyframe rule');

test(function(t) {
  var div = addDiv(t);
  div.style.animation = 'x\\yz 100s';
  assert_equals(div.getAnimations()[0].animationName, 'xyz',
                'Escaped animation name matches keyframes rule name');
}, 'Escaped animation name');

test(function(t) {
  var div = addDiv(t);
  div.style.animation = 'x\\79 z 100s';
  assert_equals(div.getAnimations()[0].animationName, 'xyz',
                'Hex-escaped animation name matches keyframes rule'
                + ' name');
}, 'Animation name with hex-escape');

</script>
</body>
